feat: add removeLike api
