+++ controllers/transaction-controller.js
@@ -1,4 +1,5 @@
 const transactionsServices = require('../services/transaction-services')
+const userServices = require('../services/user-services')
 
 const transactionsController = {
   postTransaction: (req, res, next) => {
@@ -44,6 +45,12 @@ const transactionsController = {
   },
   getPublicTransactions: (req, res, next) => {
     transactionsServices.getPublicTransactions(req, (err, data) => err ? next(err) : res.status(200).json(data))
+  },
+  addLike: (req, res, next) => {
+    transactionsServices.addLike(req, (err, data) => err ? next(err) : res.status(200).json(data))
+  },
+  removeLike: (req, res, next) => {
+    transactionsServices.removeLike(req, (err, data) => err ? next(err) : res.status(200).json(data))
   }
 }
