+++ controllers/user-controller.js
@@ -26,9 +26,6 @@ const userController = {
     const { username, introduction } = req.body
     if (!username) return res.status(400).json({ status: 'error', message: 'username is required' })
     userServices.putUser(req, { username, introduction }, (err, data) => err ? next(err) : res.status(200).json(data))
-  },
-  addLike: (req, res, next) => {
-    userServices.addLike(req, (err, data) => err ? next(err) : res.status(200).json(data))
   }
 }
